+++ src/main/java/shop/tryit/domain/order/service/OrderDetailService.java
@@ -4,9 +4,9 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import shop.tryit.domain.order.OrderDetailRepository;
 import shop.tryit.domain.order.entity.Order;
 import shop.tryit.domain.order.entity.OrderDetail;
+import shop.tryit.domain.order.repository.OrderDetailRepository;
 
 @Service
 @RequiredArgsConstructor
