+++ src/main/java/shop/tryit/domain/order/service/OrderService.java
@@ -6,11 +6,11 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import shop.tryit.domain.member.Member;
-import shop.tryit.domain.order.OrderDetailRepository;
-import shop.tryit.domain.order.OrderRepository;
 import shop.tryit.domain.order.dto.OrderSearchDto;
 import shop.tryit.domain.order.entity.Order;
 import shop.tryit.domain.order.entity.OrderDetail;
+import shop.tryit.domain.order.repository.OrderDetailRepository;
+import shop.tryit.domain.order.repository.OrderRepository;
 
 @Service
 @RequiredArgsConstructor
