+++ src/main/java/shop/tryit/repository/order/OrderDetailCustomImpl.java
@@ -1,7 +1,7 @@
 package shop.tryit.repository.order;
 
 import static shop.tryit.domain.item.entity.QItem.item;
-import static shop.tryit.domain.order.QOrderDetail.orderDetail;
+import static shop.tryit.domain.order.entity.QOrderDetail.orderDetail;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
