+++ src/test/java/shop/tryit/domain/item/ItemRepositoryTests.java
@@ -13,11 +13,11 @@ import shop.tryit.domain.item.entity.Item;
 import shop.tryit.domain.item.repository.ItemRepository;
 import shop.tryit.domain.member.Member;
 import shop.tryit.domain.member.MemberRepository;
-import shop.tryit.domain.order.OrderDetailRepository;
-import shop.tryit.domain.order.OrderRepository;
 import shop.tryit.domain.order.entity.Order;
 import shop.tryit.domain.order.entity.OrderDetail;
 import shop.tryit.domain.order.entity.OrderStatus;
+import shop.tryit.domain.order.repository.OrderDetailRepository;
+import shop.tryit.domain.order.repository.OrderRepository;
 import shop.tryit.repository.item.ItemJpaRepository;
 
 @Transactional
